+++ src/main/java/com/neconico/neconico/dto/store/card/StoreQuestionCardDto.java
@@ -1,6 +1,5 @@
-package com.neconico.neconico.vo.store;
+package com.neconico.neconico.dto.store.card;
 
-import com.neconico.neconico.vo.item.ItemCardDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.apache.ibatis.type.Alias;
@@ -20,6 +19,12 @@ public class StoreQuestionCardDto {
     private final LocalDateTime replyCreatedTime;
 
     //item
-    private ItemCardDto item;
+    private final Long itemId;
+    private final String title;
+    private final String price;
+    private final String itemImg;
+    private final LocalDateTime createdTime;
+    private final String status;
+    private final String views;
 
 }
