+++ src/main/java/com/neconico/neconico/dto/store/card/StoreReviewCardDto.java
@@ -1,4 +1,4 @@
-package com.neconico.neconico.vo.item;
+package com.neconico.neconico.dto.store.card;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,9 +8,17 @@ import java.time.LocalDateTime;
 
 @Getter
 @AllArgsConstructor
-@Alias("itemcarddto")
-public class ItemCardDto {
+@Alias("storereviewcarddto")
+//상품 문의, 후기 카드
+public class StoreReviewCardDto {
 
+    //storereview
+    private final Long reviewId;
+    private final String writerName;
+    private final String content;
+    private final LocalDateTime replyCreatedTime;
+
+    //item
     private final Long itemId;
     private final String title;
     private final String price;
