+++ src/main/java/com/neconico/neconico/service/store/StoreItemListService.java
@@ -2,10 +2,10 @@ package com.neconico.neconico.service.store;
 
 import com.neconico.neconico.dto.store.StoreItemSortingDto;
 import com.neconico.neconico.mapper.store.StoreItemListMapper;
-import com.neconico.neconico.vo.item.ItemCardDto;
-import com.neconico.neconico.vo.store.StoreQuestionCardDto;
-import com.neconico.neconico.vo.store.StoreReviewCardDto;
-import com.neconico.neconico.vo.store.StoreTradeCardDto;
+import com.neconico.neconico.dto.item.card.ItemCardDto;
+import com.neconico.neconico.dto.store.card.StoreQuestionCardDto;
+import com.neconico.neconico.dto.store.card.StoreReviewCardDto;
+import com.neconico.neconico.dto.store.card.StoreTradeCardDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
